# pass static functions to runtime (clearer intent and improves performance)