--- a/symfony/framework-bundle/5.4/public/index.php
+++ b/symfony/framework-bundle/5.4/public/index.php
@@ -4,6 +4,6 @@ use App\Kernel;
 
 require_once dirname(__DIR__).'/vendor/autoload_runtime.php';
 
-return function (array $context) {
+return static function (array $context) {
     return new Kernel($context['APP_ENV'], (bool) $context['APP_DEBUG']);
 };