| Q             | A
| ------------- | ---
| License       | MIT
| Doc issue/PR  | symfony/symfony-docs#...

<!--
Please, carefully read the README before submitting a pull request.
-->

before:
```
% autocannon -c 10 -a 100 http://127.0.0.1:8080 
Running 100 requests test @ http://127.0.0.1:8080
10 connections

┌─────────┬───────┬────────┬────────┬────────┬───────────┬──────────┬────────┐
│ Stat    │ 2.5%  │ 50%    │ 97.5%  │ 99%    │ Avg       │ Stdev    │ Max    │
├─────────┼───────┼────────┼────────┼────────┼───────────┼──────────┼────────┤
│ Latency │ 94 ms │ 442 ms │ 584 ms │ 587 ms │ 441.91 ms │ 96.86 ms │ 630 ms │
└─────────┴───────┴────────┴────────┴────────┴───────────┴──────────┴────────┘
┌───────────┬────────┬────────┬────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 14     │ 14     │ 21     │ 22      │ 20      │ 3.04   │ 14     │
├───────────┼────────┼────────┼────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 736 kB │ 736 kB │ 1.1 MB │ 1.16 MB │ 1.05 MB │ 159 kB │ 736 kB │
└───────────┴────────┴────────┴────────┴─────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.

100 requests in 5.02s, 5.25 MB read
```

after:
```
% autocannon -c 10 -a 100 http://127.0.0.1:8080
Running 100 requests test @ http://127.0.0.1:8080
10 connections

┌─────────┬────────┬────────┬────────┬────────┬──────────┬──────────┬────────┐
│ Stat    │ 2.5%   │ 50%    │ 97.5%  │ 99%    │ Avg      │ Stdev    │ Max    │
├─────────┼────────┼────────┼────────┼────────┼──────────┼──────────┼────────┤
│ Latency │ 106 ms │ 401 ms │ 472 ms │ 472 ms │ 388.1 ms │ 70.83 ms │ 473 ms │
└─────────┴────────┴────────┴────────┴────────┴──────────┴──────────┴────────┘
┌───────────┬────────┬────────┬─────────┬─────────┬─────────┬────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%     │ 97.5%   │ Avg     │ Stdev  │ Min    │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Req/Sec   │ 2      │ 2      │ 24      │ 27      │ 20      │ 9.1    │ 2      │
├───────────┼────────┼────────┼─────────┼─────────┼─────────┼────────┼────────┤
│ Bytes/Sec │ 105 kB │ 105 kB │ 1.26 MB │ 1.42 MB │ 1.05 MB │ 478 kB │ 105 kB │
└───────────┴────────┴────────┴─────────┴─────────┴─────────┴────────┴────────┘

Req/Bytes counts sampled once per second.

100 requests in 5.02s, 5.25 MB read
```