@kaznovac 2021-09-27
clearer intent is the main goal of this pr
[`static` keyword reduces the function scope](https://www.php.net/manual/en/functions.anonymous.php#functions.anonymous-functions.static) by preventing the object binding (this may help php runtime)

the benchmark shows some gains as side benefit of changing the function declaration to `static`:
on average, you can say that, overhead on the response is reduced by 10-15% (40-60ms)