@kaznovac 2021-10-08
@Nyholm thanks for the re-benchmark

regarding the meaning of `static` modifier on a closure - it clearly tells developer/reader that this closure does not require context (thus emphasizing the goal of the runtime component - decoupling from the state); potential benefits in benchmarks are side-effect (it's more to prove the change has not caused a performance regression)

regarding your results:
i find your results quite peculiar (in both runs):
- `0 2xx responses, 100000 non 2xx responses` - all non 200 responses (maybe you haven't registered default controller)
- `latency` benchmark `avg` value is multiple times less than `stdev` - deviation larger than average suggests not a normal distribution (gaussian distribution) of data (something is interfering with your throughput; e.g. garbage collector or cpu throttling, or more likely the measured time frames require measurement with nanosecond precision)

regarding statistics:
>Looking at your initial benchmark we see that the standard deviation is very high. That means** the "true" average on your "before" test is within 442 ± 96.86 ms. Compare that to the "after" 388 ± 71 ms. The both ranges overlap a lot.

both numbers for latency have got smaller, suggesting faster response and smaller variance of the response time.
you can also verify that the [null hypothesis is does not hold even for 1%](https://www.wolframalpha.com/input/?i=hypothesis+test+for+the+difference+between+two+means&assumption=%7B%22F%22%2C+%22MeanDifferenceZTest%22%2C+%22xbar1%22%7D+-%3E%22441.91%22&assumption=%7B%22F%22%2C+%22MeanDifferenceZTest%22%2C+%22mu0%22%7D+-%3E%220%22&assumption=%22FSelect%22+-%3E+%7B%7B%22MeanDifferenceZTest%22%7D%7D&assumption=%7B%22F%22%2C+%22MeanDifferenceZTest%22%2C+%22sigma1%22%7D+-%3E%2296.86%22&assumption=%7B%22F%22%2C+%22MeanDifferenceZTest%22%2C+%22sigma2%22%7D+-%3E%2270.83%22&assumption=%7B%22F%22%2C+%22MeanDifferenceZTest%22%2C+%22xbar2%22%7D+-%3E%22388.1%22&assumption=%7B%22F%22%2C+%22MeanDifferenceZTest%22%2C+%22n1%22%7D+-%3E%22100%22&assumption=%7B%22F%22%2C+%22MeanDifferenceZTest%22%2C+%22n2%22%7D+-%3E%22100%22&assumption=%7B%22F%22%2C+%22MeanDifferenceZTest%22%2C+%22c%22%7D+-%3E%220.95%22) - meaning that you have 99%+ confidence the two measurements are different.

---
allow me to state one more time: this change is about the code intent/clarity not about performance