@Nyholm 2021-10-08
Oh. Sorry. you are correct. My results are invalid =). I used http instead of https. Silly me. 

Here are the new results: 

### Before 100
```
❯ autocannon -c 10 -a 100 https://127.0.0.1:8000
Running 100 requests test @ https://127.0.0.1:8000
10 connections

┌─────────┬───────┬───────┬────────┬────────┬──────────┬──────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5%  │ 99%    │ Avg      │ Stdev    │ Max    │
├─────────┼───────┼───────┼────────┼────────┼──────────┼──────────┼────────┤
│ Latency │ 32 ms │ 46 ms │ 113 ms │ 122 ms │ 51.52 ms │ 19.51 ms │ 122 ms │
└─────────┴───────┴───────┴────────┴────────┴──────────┴──────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬───────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼───────┼─────────┤
│ Req/Sec   │ 100     │ 100     │ 100     │ 100     │ 100     │ 0     │ 100     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼───────┼─────────┤
│ Bytes/Sec │ 18.6 kB │ 18.6 kB │ 18.6 kB │ 18.6 kB │ 18.6 kB │ 0 B   │ 18.6 kB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴───────┴─────────┘

Req/Bytes counts sampled once per second.

100 requests in 1.03s, 18.6 kB read
```

### After 100
```
❯ autocannon -c 10 -a 100 https://127.0.0.1:8000
Running 100 requests test @ https://127.0.0.1:8000
10 connections

┌─────────┬───────┬───────┬────────┬────────┬──────────┬──────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5%  │ 99%    │ Avg      │ Stdev    │ Max    │
├─────────┼───────┼───────┼────────┼────────┼──────────┼──────────┼────────┤
│ Latency │ 34 ms │ 45 ms │ 112 ms │ 113 ms │ 50.29 ms │ 19.22 ms │ 119 ms │
└─────────┴───────┴───────┴────────┴────────┴──────────┴──────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬───────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼───────┼─────────┤
│ Req/Sec   │ 100     │ 100     │ 100     │ 100     │ 100     │ 0     │ 100     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼───────┼─────────┤
│ Bytes/Sec │ 18.6 kB │ 18.6 kB │ 18.6 kB │ 18.6 kB │ 18.6 kB │ 0 B   │ 18.6 kB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴───────┴─────────┘

Req/Bytes counts sampled once per second.

100 requests in 1.02s, 18.6 kB read
```

### Before 10.000
```
❯ autocannon -c 10 -a 10000 https://127.0.0.1:8000
Running 10000 requests test @ https://127.0.0.1:8000
10 connections

┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
│ Latency │ 34 ms │ 43 ms │ 56 ms │ 61 ms │ 43.29 ms │ 6.09 ms │ 118 ms │
└─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
┌───────────┬─────────┬───────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%  │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼───────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 190     │ 204   │ 229     │ 239     │ 227.28  │ 9.88    │ 190     │
├───────────┼─────────┼───────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 35.4 kB │ 38 kB │ 42.6 kB │ 44.5 kB │ 42.3 kB │ 1.84 kB │ 35.3 kB │
└───────────┴─────────┴───────┴─────────┴─────────┴─────────┴─────────┴─────────┘

Req/Bytes counts sampled once per second.

10k requests in 44.04s, 1.86 MB read
```

### After 10.000
```
❯ autocannon -c 10 -a 10000 https://127.0.0.1:8000
Running 10000 requests test @ https://127.0.0.1:8000
10 connections

┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬────────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max    │
├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼────────┤
│ Latency │ 35 ms │ 44 ms │ 61 ms │ 69 ms │ 45.09 ms │ 7.31 ms │ 140 ms │
└─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴────────┘
┌───────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┬─────────┐
│ Stat      │ 1%      │ 2.5%    │ 50%     │ 97.5%   │ Avg     │ Stdev   │ Min     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Req/Sec   │ 143     │ 180     │ 222     │ 233     │ 217.4   │ 16.17   │ 143     │
├───────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┼─────────┤
│ Bytes/Sec │ 26.6 kB │ 33.5 kB │ 41.3 kB │ 43.4 kB │ 40.4 kB │ 3.01 kB │ 26.6 kB │
└───────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┴─────────┘

Req/Bytes counts sampled once per second.

10k requests in 46.04s, 1.86 MB read
```

The performance says: It is pretty much the same. 

-------

> regarding the meaning of static modifier on a closure - it clearly tells developer/reader that this closure does not require context (thus emphasizing the goal of the runtime component - decoupling from the state); potential benefits in benchmarks are side-effect (it's more to prove the change has not caused a performance regression)

> allow me to state one more time: this change is about the code intent/clarity not about performance

Noted that this is about code intent and not performance.

From the [manual](https://www.php.net/manual/en/functions.anonymous.php#functions.anonymous-functions.static): 

> Anonymous functions may be declared statically. This prevents them from having the current class automatically bound to them. Objects may also not be bound to them at runtime.

Since this function is not declared in a class, it is already static. 
This PR will make sure one cannot bind the closure to a class at runtime, but it makes no difference as [pointed out before](https://github.com/symfony/recipes/pull/984#issuecomment-901081438). 

So the way I see it, this PR (and #984) is just about code style. Please correct me if Im wrong.